requirements.txt
smbus2>=0.4.0
pyserial>=3.5
pynmea2>=1.19.0
python-obd>=0.7.1
